Re-consolidated ReadFrames and WriteFrames into DataFrame. Introduced BoundedInputStream. Introduced DataFrameInputStream: completely eliminated the read double-buffering.
