+++ segmentstore/server/src/test/java/io/pravega/segmentstore/server/logs/DataFrameBuilderTests.java
@@ -119,7 +119,8 @@ public class DataFrameBuilderTests extends ThreadPooledTestSuite {
             // Read all entries in the Log and interpret them as DataFrames, then verify the records can be reconstructed.
             await(() -> commitFrames.size() >= order.size(), 20);
 
-            List<ReadFrame> frames = dataLog.getAllEntries(readItem -> new ReadFrame(readItem.getPayload(), readItem.getLength()));
+            List<DataFrame.DataFrameEntryIterator> frames = dataLog.getAllEntries(readItem ->
+                    DataFrame.read(readItem.getPayload(), readItem.getLength(), readItem.getAddress()));
             Assert.assertEquals("Unexpected number of frames generated.", commitFrames.size(), frames.size());
 
             // Check the correctness of the commit callback.
@@ -208,7 +209,7 @@ public class DataFrameBuilderTests extends ThreadPooledTestSuite {
         AssertExtensions.assertListEquals("Items read back do not match expected values.", expectedItems, readItems, TestLogItem::equals);
 
         // Read all entries in the Log and interpret them as DataFrames, then verify the records can be reconstructed.
-        List<ReadFrame> frames = dataLog.getAllEntries(ri -> new ReadFrame(ri.getPayload(), ri.getLength()));
+        val frames = dataLog.getAllEntries(ri -> DataFrame.read(ri.getPayload(), ri.getLength(), ri.getAddress()));
 
         // Check the correctness of the commit callback.
         AssertExtensions.assertGreaterThan("Not enough Data Frames were generated.", 1, frames.size());
@@ -255,7 +256,7 @@ public class DataFrameBuilderTests extends ThreadPooledTestSuite {
             Assert.assertEquals("Exactly one Data Frame was expected so far.", 1, commitFrames.size());
 
             //Read all entries in the Log and interpret them as DataFrames, then verify the records can be reconstructed.
-            List<ReadFrame> frames = dataLog.getAllEntries(readItem -> new ReadFrame(readItem.getPayload(), readItem.getLength()));
+            val frames = dataLog.getAllEntries(readItem -> DataFrame.read(readItem.getPayload(), readItem.getLength(), readItem.getAddress()));
             Assert.assertEquals("Unexpected number of frames generated.", commitFrames.size(), frames.size());
             DataFrameTestHelpers.checkReadRecords(frames, records, r -> new ByteArraySegment(r.getFullSerialization()));
         }
@@ -309,7 +310,7 @@ public class DataFrameBuilderTests extends ThreadPooledTestSuite {
             }
 
             //Read all entries in the Log and interpret them as DataFrames, then verify the records can be reconstructed.
-            List<ReadFrame> frames = dataLog.getAllEntries(readItem -> new ReadFrame(readItem.getPayload(), readItem.getLength()));
+            val frames = dataLog.getAllEntries(readItem -> DataFrame.read(readItem.getPayload(), readItem.getLength(), readItem.getAddress()));
             DataFrameTestHelpers.checkReadRecords(frames, records, r -> new ByteArraySegment(r.getFullSerialization()));
         }
     }
